Update Webdriver.py
I have modified "createWebdriver" to eliminate the need to specify the platform when "isDockerized" is "True." This change was made to prevent confusion in the "config.yaml" file. In summary, it is no longer necessary to include "platform: "linux"" for the Dockerized version to obtain the correct path for the chromedriver.