Updated content-type check in the condition to handle both "application/json" and "text/plain" packet types.
This change addresses the issue where 'watchHour' is not populated.

"/stats?sport=lol" content-type changed to text/plain for some reason. This caused 'watchHour' to not be populated as expected. By updating the content-type check in the condition to include both "application/json" and "text/plain" packet types, this issue is resolved.

Affected file: NetworkHandler.py